Update: Separate buttons to show and hide passwords as component
