+++ pages/login/index.module.css
@@ -16,7 +16,7 @@
 .title {
   color: var(--darkgreen);
   text-align: center;
-  padding: 1em 0;
+  padding: 1em 0 0.5em;
   font-size: 1.8em;
 }
 
@@ -35,12 +35,12 @@
   border: 1px solid var(--lightgray);
   border-radius: 5px;
   padding: 10px 15px;
-  margin-bottom: 15px;
+  margin-top: 15px;
   width: 100%;
 }
 
 .formBtn {
-  margin: 0 0 15px;
+  margin: 25px 0;
   padding: 10px 15px;
   width: 50%;
   border: none;
@@ -69,18 +69,6 @@
   font-size: 0.9rem;
   text-align: left;
   margin: 0 10px 20px;
-}
-
-.passwordReveal {
-  left: auto;
-  position: absolute;
-  right: 10px;
-  top: 14px;
-}
-
-.passwordIcon {
-  font-size: 1.2rem;
-  color: var(--gray);
 }
 
 @media screen and (max-width: 640px) {
