+++ pages/reset-password/index.js
@@ -1,8 +1,10 @@
-import { useState } from "react";
+import { useContext, useState } from "react";
 import { useRouter } from "next/router";
 import styles from "./index.module.css";
 import { resetPwd } from "../../lib/auth";
 import Header from "../../components/Header";
+import TogglePassword from "@/components/TogglePassword";
+import AppContext from "@/context/AppContext";
 
 const ResetPwd = () => {
   const router = useRouter();
@@ -10,6 +12,7 @@ const ResetPwd = () => {
     password: "",
     passwordConfirmation: "",
   });
+  const { passwordType } = useContext(AppContext);
 
   const handleResetPwd = () => {
     const { code } = router.query;
@@ -27,28 +30,29 @@ const ResetPwd = () => {
       <Header />
       <div className={styles.formWrapper}>
         <h1 className={styles.title}>Reset password</h1>
-        <p className={styles.message}>Please set a new password.</p>
         <div className={styles.form}>
           <div className={styles.formItem}>
             <input
               className={styles.formInput}
-              type="password"
+              type={passwordType}
               id="password"
               name="password"
               placeholder="Password"
               autoFocus={true}
               onChange={handleChange}
             />
+            <TogglePassword />
           </div>
           <div className={styles.formItem}>
             <input
               className={styles.formInput}
-              type="password"
+              type={passwordType}
               id="passwordConfirmation"
               name="passwordConfirmation"
               placeholder="Password confirmation"
               onChange={handleChange}
             />
+            <TogglePassword />
           </div>
           <button className={styles.formBtn} onClick={handleResetPwd}>
             Apply
